change positioning from absolute to relative
with absolute, the body's width was wrong, it was less than half the amount it should be. this was because most of the objects were overflowing outside of it. changing it to relative and removing the top positioning fixes this and makes the body have the full height of th page which makes it much easier to add a footer.